Add solution for 2018 Day 10.
This one was a odd solutionas you needed to read the letters plotted on an image.
I couldn't get it to write to the console (the boundaries were too large)
so it outputs bitmaps to a hard-coded filepath to be manually checked.

The answer for part 2 is simply the image name - ie. the number of steps (seconds) taken